✨ Adding the equation steps function
